[web] Rename Overview to OverviewPage
